+++ src/frontend/src/features/rooms/api/muteParticipant.ts
@@ -1,12 +1,11 @@
 import { Participant, Track } from 'livekit-client'
 import Source = Track.Source
-import { fetchServerApi } from './fetchServerApi'
-import { buildServerApiUrl } from './buildServerApiUrl'
-import { useRoomData } from '../hooks/useRoomData'
+import { useRoomData } from '../livekit/hooks/useRoomData'
 import {
   useNotifyParticipants,
   NotificationType,
 } from '@/features/notifications'
+import { fetchApi } from '@/api/fetchApi'
 
 export const useMuteParticipant = () => {
   const data = useRoomData()
@@ -14,8 +13,8 @@ export const useMuteParticipant = () => {
   const { notifyParticipants } = useNotifyParticipants()
 
   const muteParticipant = async (participant: Participant) => {
-    if (!data || !data?.livekit) {
-      throw new Error('Room data is not available')
+    if (!data?.id) {
+      throw new Error('Room id is not available')
     }
     const trackSid = participant.getTrackPublication(
       Source.Microphone
@@ -26,22 +25,13 @@ export const useMuteParticipant = () => {
     }
 
     try {
-      const response = await fetchServerApi(
-        buildServerApiUrl(
-          data.livekit.url,
-          'twirp/livekit.RoomService/MutePublishedTrack'
-        ),
-        data.livekit.token,
-        {
-          method: 'POST',
-          body: JSON.stringify({
-            room: data.livekit.room,
-            identity: participant.identity,
-            muted: true,
-            track_sid: trackSid,
-          }),
-        }
-      )
+      const response = await fetchApi(`rooms/${data.id}/mute-participant/`, {
+        method: 'POST',
+        body: JSON.stringify({
+          participant_identity: participant.identity,
+          track_sid: trackSid,
+        }),
+      })
 
       await notifyParticipants({
         type: NotificationType.ParticipantMuted,
